+++ src/decision.c
@@ -22,7 +22,7 @@
 #include "dcodegen.h"
 #include "derror.h"
 #include "dlex.h"
-#include "dlink.h"  
+#include "dlink.h"
 #include "dmalloc.h"
 #include "dobj.h"
 #include "doptimize.h"
@@ -159,7 +159,7 @@ static const char *load_string_from_file(const char *filePath, size_t *size,
 static void save_object_to_file(const char *filePath, const char *content,
                                 size_t len) {
     FILE *f;
-    
+
     f = fopen(filePath, "wb");
     if (f == NULL) {
         printf("Can't open the file!\n");
@@ -185,7 +185,7 @@ bool d_run_sheet(Sheet *sheet) {
         if (sheet->_isLinked) {
             if (sheet->_main > 0) // A Start function exists.
             {
-                DVM vm = d_vm_create();
+                DVM vm       = d_vm_create();
                 bool success = d_vm_run(&vm, sheet->_text + sheet->_main);
                 d_vm_free(&vm);
                 return success;
@@ -207,9 +207,9 @@ bool d_run_sheet(Sheet *sheet) {
  * \fn bool d_run_function(DVM *vm, Sheet *sheet, const char *funcName)
  * \brief Run the specified function/subroutine in a given sheet, given the
  * sheet has gone through `d_codegen_compile`.
- * 
+ *
  * \return If the function/subroutine ran without any errors.
- * 
+ *
  * \param vm The VM to run the function on. The reason it is a seperate
  * argument is because it allows you to push and pop arguments and return values
  * seperately.
@@ -290,7 +290,8 @@ bool d_run_function(DVM *vm, Sheet *sheet, const char *funcName) {
 }
 
 /**
- * \fn Sheet *d_load_string(const char *source, const char *name)
+ * \fn Sheet *d_load_string(const char *source, const char *name,
+ *                          Sheet **includes)
  * \brief Take Decision source code and compile it into bytecode, but do not
  * run it.
  *
@@ -298,13 +299,24 @@ bool d_run_function(DVM *vm, Sheet *sheet, const char *funcName) {
  *
  * \param source The source code to compile.
  * \param name The name of the sheet. If NULL, it is set to `"source"`.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-Sheet *d_load_string(const char *source, const char *name) {
+Sheet *d_load_string(const char *source, const char *name, Sheet **includes) {
     // If name is NULL, set a name.
     name = (name != NULL) ? name : "source";
 
     // Represent the sheet in memory, and check for errors along the way.
     Sheet *sheet = d_sheet_create(name);
+
+    if (includes != NULL) {
+        Sheet **include = includes;
+
+        while (*include) {
+            d_sheet_add_include(sheet, *include);
+            include++;
+        }
+    }
 
     VERBOSE(1, "--- STAGE 1: Creating lexical stream...\n")
     LexStream stream = d_lex_create_stream(source, name);
@@ -364,7 +376,7 @@ Sheet *d_load_string(const char *source, const char *name) {
 }
 
 /**
- * \fn bool d_run_string(const char *source, const char *name)
+ * \fn bool d_run_string(const char *source, const char *name, Sheet **includes)
  * \brief Take Decision source code and compile it into bytecode. If it
  * compiled successfully, run it in the virtual machine.
  *
@@ -372,9 +384,11 @@ Sheet *d_load_string(const char *source, const char *name) {
  *
  * \param source The source code the compile.
  * \param name The name of the sheet. If `NULL`, it is set to `"source"`.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_run_string(const char *source, const char *name) {
-    Sheet *sheet   = d_load_string(source, name);
+bool d_run_string(const char *source, const char *name, Sheet **includes) {
+    Sheet *sheet   = d_load_string(source, name, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -390,7 +404,8 @@ bool d_run_string(const char *source, const char *name) {
 }
 
 /**
- * \fn bool d_compile_string(const char *source, const char *filePath)
+ * \fn bool d_compile_string(const char *source, const char *filePath,
+ *                           Sheet **includes)
  * \brief Take Decision source code and compile it into bytecode. Then save
  * it into a binary file if it compiled successfully.
  *
@@ -398,9 +413,12 @@ bool d_run_string(const char *source, const char *name) {
  *
  * \param source The source code to compile.
  * \param filePath Where to write the object file to.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_compile_string(const char *source, const char *filePath) {
-    Sheet *sheet   = d_load_string(source, NULL);
+bool d_compile_string(const char *source, const char *filePath,
+                      Sheet **includes) {
+    Sheet *sheet   = d_load_string(source, NULL, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -417,21 +435,23 @@ bool d_compile_string(const char *source, const char *filePath) {
 }
 
 /**
- * \fn Sheet *d_load_source_file(const char *filePath)
+ * \fn Sheet *d_load_source_file(const char *filePath, Sheet **includes)
  * \brief Take Decision source code from a file and compile it into bytecode,
  * but do not run it.
  *
  * \return A malloc'd sheet containing all of the compilation info.
  *
  * \param filePath The file path of the source file to compile.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-Sheet *d_load_source_file(const char *filePath) {
+Sheet *d_load_source_file(const char *filePath, Sheet **includes) {
     size_t _size; // Not needed.
     const char *source = load_string_from_file(filePath, &_size, false);
     Sheet *sheet       = NULL;
 
     if (source != NULL) {
-        sheet = d_load_string(source, filePath);
+        sheet = d_load_string(source, filePath, includes);
         free((void *)source);
     } else {
         // We errored loading the file.
@@ -443,16 +463,18 @@ Sheet *d_load_source_file(const char *filePath) {
 }
 
 /**
- * \fn bool d_run_source_file(const char *filePath)
+ * \fn bool d_run_source_file(const char *filePath, Sheet **includes)
  * \brief Take Decision source code in a file and compile it into bytecode. If
  * it compiled successfully, run it in the virtual machine.
  *
  * \return If the code compiled/ran without any errors.
  *
  * \param filePath The file path of the source file to compile.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_run_source_file(const char *filePath) {
-    Sheet *sheet   = d_load_source_file(filePath);
+bool d_run_source_file(const char *filePath, Sheet **includes) {
+    Sheet *sheet   = d_load_source_file(filePath, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -468,7 +490,8 @@ bool d_run_source_file(const char *filePath) {
 }
 
 /**
- * \fn bool d_compile_file(const char *filePathIn, const char *filePathOut)
+ * \fn bool d_compile_file(const char *filePathIn, const char *filePathOut
+ *                         Sheet **includes)
  * \brief Take Decision source code from a file and compile it into bytecode.
  * If it compiled successfully, save it into a binary file.
  *
@@ -476,9 +499,12 @@ bool d_run_source_file(const char *filePath) {
  *
  * \param filePathIn The file path of the source file to compile.
  * \param filePathOut Where to write the object file to.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_compile_file(const char *filePathIn, const char *filePathOut) {
-    Sheet *sheet   = d_load_source_file(filePathIn);
+bool d_compile_file(const char *filePathIn, const char *filePathOut,
+                    Sheet **includes) {
+    Sheet *sheet   = d_load_source_file(filePathIn, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -495,20 +521,22 @@ bool d_compile_file(const char *filePathIn, const char *filePathOut) {
 }
 
 /**
- * \fn Sheet *d_load_object_file(const char *filePath)
+ * \fn Sheet *d_load_object_file(const char *filePath, Sheet **includes)
  * \brief Take a Decision object file and load it into memory.
  *
  * \return A malloc'd sheet object containing all of the compilation info.
  *
  * \param filePath The file path of the object file.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-Sheet *d_load_object_file(const char *filePath) {
+Sheet *d_load_object_file(const char *filePath, Sheet **includes) {
     size_t size;
     const char *obj = load_string_from_file(filePath, &size, true);
     Sheet *out      = NULL;
 
     if (obj != NULL) {
-        out = d_obj_load(obj, size, filePath);
+        out = d_obj_load(obj, size, filePath, includes);
         free((char *)obj);
 
         out->hasErrors = d_error_report();
@@ -526,16 +554,18 @@ Sheet *d_load_object_file(const char *filePath) {
 }
 
 /**
- * \fn bool d_run_object_file(const char *filePath)
+ * \fn bool d_run_object_file(const char *filePath, Sheet **includes)
  * \brief Take a Decision object file, load it into memory, and run it in the
  * virtual machine.
  *
  * \return If the code ran without any errors.
  *
  * \param filePath The file path of the object file.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_run_object_file(const char *filePath) {
-    Sheet *sheet   = d_load_object_file(filePath);
+bool d_run_object_file(const char *filePath, Sheet **includes) {
+    Sheet *sheet   = d_load_object_file(filePath, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -584,15 +614,17 @@ short d_is_object_file(const char *filePath) {
 }
 
 /**
- * \fn Sheet *d_load_file(const char *filePath)
+ * \fn Sheet *d_load_file(const char *filePath, Sheet **includes)
  * \brief Take a Decision file, decide whether it is a source or an object file
  * based on its contents, and load it into memory.
  *
  * \return A malloc'd sheet object containing all of the compilation info.
  *
  * \param filePath The file path of the file to load.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-Sheet *d_load_file(const char *filePath) {
+Sheet *d_load_file(const char *filePath, Sheet **includes) {
     short fileType = d_is_object_file(filePath);
 
     Sheet *out;
@@ -600,13 +632,13 @@ Sheet *d_load_file(const char *filePath) {
     switch (fileType) {
         case 0:
             // It is a source file.
-            out = d_load_source_file(filePath);
+            out = d_load_source_file(filePath, includes);
 
             break;
 
         case 1:
             // It is an object file.
-            out = d_load_object_file(filePath);
+            out = d_load_object_file(filePath, includes);
 
             break;
 
@@ -622,16 +654,18 @@ Sheet *d_load_file(const char *filePath) {
 }
 
 /**
- * \fn bool d_run_file(const char *filePath)
+ * \fn bool d_run_file(const char *filePath, Sheet **includes)
  * \brief Take a Decision file, decide whether it is a source or an object file
  * based on its contents, and run it in the virtual machine.
  *
  * \return If the code compiled/ran without any errors.
  *
  * \param filePath The file path of the file to load.
+ * \param includes A NULL-terminated list of initially included sheets.
+ * Can be NULL.
  */
-bool d_run_file(const char *filePath) {
-    Sheet *sheet   = d_load_file(filePath);
+bool d_run_file(const char *filePath, Sheet **includes) {
+    Sheet *sheet   = d_load_file(filePath, includes);
     bool hadErrors = sheet->hasErrors;
 
     if (!hadErrors) {
@@ -760,7 +794,7 @@ int main(int argc, char *argv[]) {
 
                 // Now we have our new objFilePath, let's compile!
                 d_compile_file((const char *)filePath,
-                               (const char *)objFilePath);
+                               (const char *)objFilePath, NULL);
 
                 free(objFilePath);
             }
@@ -779,13 +813,13 @@ int main(int argc, char *argv[]) {
 
             if (isObjectFile) {
                 if (disassemble) {
-                    Sheet *sheet = d_load_object_file((const char *)filePath);
+                    Sheet *sheet = d_load_object_file((const char *)filePath, NULL);
                     d_asm_dump_all(sheet);
                     d_sheet_free(sheet);
 
                     return 0;
                 } else
-                    return d_run_object_file((const char *)filePath);
+                    return d_run_object_file((const char *)filePath, NULL);
             } else {
                 // Check that we are not disassembling a source file.
                 if (disassemble) {
@@ -793,7 +827,7 @@ int main(int argc, char *argv[]) {
                            "object file!\n");
                     return 1;
                 } else
-                    return d_run_source_file((const char *)filePath);
+                    return d_run_source_file((const char *)filePath, NULL);
             }
         }
     } else {