+++ tests/c/c_from_decision.c
@@ -71,25 +71,41 @@ void myFactorial(DVM *vm) {
 }
 
 int main() {
-    DType halfInputs[]  = {TYPE_FLOAT, TYPE_NONE};
-    DType halfOutputs[] = {TYPE_FLOAT, TYPE_NONE};
 
-    // d_create_c_function
-    d_create_c_function("Half", &myHalf, halfInputs, halfOutputs);
+    SocketMeta halfSockets[] = {
+        {"number", "The number to half", TYPE_FLOAT, {0}},
+        {"half", "The number halved", TYPE_FLOAT, {0}}};
 
-    DType canDriveInputs[]  = {TYPE_INT, TYPE_BOOL, TYPE_NONE};
-    DType canDriveOutputs[] = {TYPE_BOOL, TYPE_STRING, TYPE_NONE};
+    CFunction halfFunction = d_create_c_function(
+        &myHalf, "Half", "Half a number.", halfSockets, 1, 1);
 
-    // d_create_c_function
-    d_create_c_function("CanDrive", &myCanDrive, canDriveInputs,
-                        canDriveOutputs);
+    SocketMeta canDriveSockets[] = {
+        {"age", "The age of the person.", TYPE_INT, {0}},
+        {"hasLicense", "Has the person got a driving license?", TYPE_BOOL, {0}},
+        {"canDrive", "If the person can drive.", TYPE_BOOL, {0}},
+        {"reason", "If the person cannot drive, why?", TYPE_STRING, {0}}};
 
-    DType factorialInputs[]  = {TYPE_INT, TYPE_NONE};
-    DType factorialOutputs[] = {TYPE_INT, TYPE_NONE};
+    CFunction canDriveFunction = d_create_c_function(
+        &myCanDrive, "CanDrive", "Determine if a person can drive.",
+        canDriveSockets, 2, 2);
 
-    // d_create_c_subroutine
-    d_create_c_subroutine("Factorial", &myFactorial, factorialInputs,
-                          factorialOutputs);
+    SocketMeta factorialSockets[] = {
+        {"n", "The number to get the factorial of.", TYPE_INT, {0}},
+        {"nFactorial", "The factorial of n, n!", TYPE_INT, {0}}};
+
+    CFunction factorialSubroutine = d_create_c_subroutine(
+        &myFactorial, "Factorial", "Calculate the factorial of an integer.",
+        factorialSockets, 1, 1);
+
+    Sheet *library     = d_sheet_create("MyFunctions");
+    library->allowFree = false;
+
+    d_sheet_add_c_function(library, halfFunction);
+    d_sheet_add_c_function(library, canDriveFunction);
+    d_sheet_add_c_function(library, factorialSubroutine);
+
+    Sheet *includeList[] = {NULL, NULL};
+    includeList[0]       = library;
 
     char *src = "Start~#1\n"
                 "Half(100.125)~#2\n"
@@ -98,13 +114,13 @@ int main() {
                 "Print(#3, #4)~#6\n"
                 "Print(#6, #5)~#7\n"
                 "Factorial(#7, 10)~#8, #9\n"
-                "Print(#8, #9)\n";
+                "Print(#9, #8)\n";
 
     char *answer = "50.0625\nfalse\nYou do not have a license.\n3628800\n";
-    
+
     // d_run_string
     START_CAPTURE_STDOUT()
-    d_run_string(src, NULL);
+    d_run_string(src, NULL, includeList);
     STOP_CAPTURE_STDOUT()
     ASSERT_CAPTURED_STDOUT(answer)
 
@@ -121,20 +137,20 @@ int main() {
 
     // d_run_file
     START_CAPTURE_STDOUT()
-    d_run_file("main.dc");
+    d_run_file("main.dc", includeList);
     STOP_CAPTURE_STDOUT()
     ASSERT_CAPTURED_STDOUT(answer)
 
     // d_compile_string
-    d_compile_string(src, "main.dco");
+    d_compile_string(src, "main.dco", includeList);
 
     // d_run_file
     START_CAPTURE_STDOUT()
-    d_run_file("main.dco");
+    d_run_file("main.dco", includeList);
     STOP_CAPTURE_STDOUT()
     ASSERT_CAPTURED_STDOUT(answer)
-    
-    d_free_c_functions();
+
+    d_sheet_free(library);
 
     return 0;
 }