+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and descriptions which are accessible via the C API.
   * Finding the definitions of names has been moved to `dname.c` and `dname.h`.
   * Decision object files now have their own set of source files, `dobj.c` and
   `dobj.h`, and the code has been revamped to be less error-prone.
+* C functions are now stored in sheets, rather than in a global list.
+* Sheets now have an `allowFree` property that stops sheets that include it
+from freeing the sheet when they are freed.
 
 # Decision v0.2.0 - 16th January 2020